Update README.md
Introduction to Truffle and ERC20 Projects

Truffle is a popular development framework for Ethereum that facilitates the building, testing, and deployment of decentralized applications (dApps). One of the common tasks in Ethereum development is creating and managing ERC20 tokens, which are standard tokens used on the Ethereum blockchain that allow for the transfer and management of cryptocurrency. An ERC20 project using Truffle involves initializing the project, creating the token contract, and deploying it to the Ethereum network.